example/voiceAgent/requirements.txt
websockets>=13.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
requests>=2.31.0

